BTreeMap: split off most code of remove and split_off
